# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "APIError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "description": "Stable machine-readable code.",
            "type": "string"
          },
          "errors": {
            "description": "Per-field problems, when applicable.",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array"
          },
          "message": {
            "description": "Human-readable explanation.",
            "type": "string"
          },
          "status": {
            "description": "HTTP status code.",
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "status",
          "code",
          "message"
        ],
        "type": "object"
      },
...
      "Device": {
        "additionalProperties": false,
        "properties": {
          "app_version_code": {
            "format": "int32",
            "type": [
              "integer",
              "null"
            ]
          },
          "app_version_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "brand": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "heartbeat_interval_minutes": {
            "format": "int32",
            "type": "integer"
          },
          "id": {
            "type": "string"
          },
          "is_default": {
            "type": "boolean"
          },
          "last_heartbeat_at": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "manufacturer": {
            "type": [
              "string",
              "null"
            ]
          },
          "model": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "online": {
            "type": "boolean"
          },
          "os_api_level": {
            "format": "int32",
            "type": [
              "integer",
              "null"
            ]
          },
          "os_version": {
            "type": [
              "string",
              "null"
            ]
          },
          "preferred_sim_subscription_id": {
            "format": "int32",
            "type": [
              "integer",
              "null"
            ]
          },
          "push_token_invalid_reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "push_token_invalidated_at": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "receive_enabled": {
            "type": "boolean"
          },
          "received_count": {
            "format": "int64",
            "type": "integer"
          },
          "send_delay_seconds": {
            "format": "int32",
            "type": "integer"
          },
          "sent_count": {
            "format": "int64",
            "type": "integer"
          },
          "sims": {},
          "telemetry": {},
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "enabled",
          "is_default",
          "manufacturer",
          "brand",
          "model",
          "os_version",
          "os_api_level",
          "app_version_name",
          "app_version_code",
          "push_token_invalidated_at",
          "push_token_invalid_reason",
          "receive_enabled",
          "send_delay_seconds",
          "heartbeat_interval_minutes",
          "preferred_sim_subscription_id",
          "last_heartbeat_at",
          "online",
          "telemetry",
          "sims",
          "sent_count",
          "received_count",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      },
...
      "FieldError": {
        "additionalProperties": false,
        "properties": {
          "field": {
            "description": "Location of the problem, for example body.to[0].",
            "type": "string"
          },
          "message": {
            "description": "What is wrong with it.",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "type": "object"
      },
...
      "LoginInputBody": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "format": "email",
            "maxLength": 254,
            "type": "string"
          },
          "password": {
            "maxLength": 200,
            "type": "string"
          },
          "turnstile_token": {
            "description": "Cloudflare Turnstile token. Required when GET /v1/auth/config reports a turnstile_site_key.",
            "maxLength": 2048,
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "type": "object"
      },
...
      "PairInputBody": {
        "additionalProperties": false,
        "properties": {
          "app_version_code": {
            "format": "int32",
            "type": "integer"
          },
          "app_version_name": {
            "maxLength": 32,
            "type": "string"
          },
          "brand": {
            "maxLength": 64,
            "type": "string"
          },
          "build_id": {
            "maxLength": 64,
            "type": "string"
          },
          "code": {
            "description": "Pairing code from the dashboard.",
            "maxLength": 9,
            "minLength": 8,
            "type": "string"
          },
          "hardware_key": {
            "description": "Stable hash the app derives from the handset, so re-pairing finds the same device.",
            "maxLength": 128,
            "minLength": 8,
            "type": "string"
          },
          "manufacturer": {
            "maxLength": 64,
            "type": "string"
          },
          "model": {
            "maxLength": 64,
            "type": "string"
          },
          "name": {
            "maxLength": 64,
            "type": "string"
          },
          "os_api_level": {
            "format": "int32",
            "type": "integer"
          },
          "os_version": {
            "maxLength": 32,
            "type": "string"
          },
          "push_token": {
            "maxLength": 4096,
            "type": "string"
          }
        },
        "required": [
          "code",
          "hardware_key"
        ],
        "type": "object"
      },
      "PairOutputBody": {
        "additionalProperties": false,
        "properties": {
          "device": {
            "$ref": "#/components/schemas/Device"
          },
          "device_token": {
            "description": "Bearer token for /v1/device endpoints. Shown once.",
            "type": "string"
          }
        },
        "required": [
          "device",
          "device_token"
        ],
        "type": "object"
      },
...
      "RegisterInputBody": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "description": "Sign-in email. A verification code is sent to it.",
            "format": "email",
            "maxLength": 254,
            "type": "string"
          },
          "name": {
            "description": "Display name.",
            "maxLength": 100,
            "type": "string"
          },
          "password": {
            "description": "At least 10 characters.",
            "maxLength": 200,
            "minLength": 10,
            "type": "string"
          },
          "turnstile_token": {
            "description": "Cloudflare Turnstile token. Required when GET /v1/auth/config reports a turnstile_site_key.",
            "maxLength": 2048,
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "type": "object"
      },
...
      "ResetInputBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "maxLength": 6,
            "minLength": 6,
            "type": "string"
          },
          "email": {
            "format": "email",
            "maxLength": 254,
            "type": "string"
          },
          "new_password": {
            "maxLength": 200,
            "minLength": 10,
            "type": "string"
          }
        },
        "required": [
          "email",
          "code",
          "new_password"
        ],
        "type": "object"
      },
      "ResetRequestInputBody": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "format": "email",
            "maxLength": 254,
            "type": "string"
          },
          "turnstile_token": {
            "description": "Cloudflare Turnstile token. Required when GET /v1/auth/config reports a turnstile_site_key.",
            "maxLength": 2048,
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "type": "object"
      },
...
      "SessionOutputBody": {
        "additionalProperties": false,
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "user"
        ],
        "type": "object"
      },
...
      "User": {
        "additionalProperties": false,
        "properties": {
          "avatar_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "deletion_requested_at": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "email": {
            "type": "string"
          },
          "email_verified_at": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "google_linked": {
            "type": "boolean"
          },
          "has_password": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "last_login_at": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "onboarding": {},
          "role": {
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "avatar_url",
          "role",
          "email_verified_at",
          "deletion_requested_at",
          "onboarding",
          "last_login_at",
          "created_at",
          "updated_at",
          "has_password",
          "google_linked"
        ],
        "type": "object"
      },
      "UserOutputBody": {
        "additionalProperties": false,
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "user"
        ],
        "type": "object"
      },
      "VerifyEmailInputBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "description": "The 6-digit code from the email.",
            "maxLength": 6,
            "minLength": 6,
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "description": "Developer API key from the dashboard.",
        "in": "header",
        "name": "X-Api-Key",
        "type": "apiKey"
      },
      "deviceToken": {
        "description": "Device token issued at pairing. Only valid on /v1/device endpoints.",
        "scheme": "bearer",
        "type": "http"
      },
      "session": {
        "description": "Dashboard session cookie.",
        "in": "cookie",
        "name": "simhook_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Send and receive SMS through your own Android phone. Authenticate with an API key in the X-Api-Key header.",
    "title": "simhook API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/auth/login": {
      "post": {
        "description": "Checks the password and sets the session cookies the dashboard uses. A session lives 30 days without use, longer while it is used, and 180 days at most.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOutputBody"
                }
              }
            },
            "description": "OK",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Sign in",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/logout": {
      "post": {
        "description": "Ends this session and clears the cookies. Always succeeds, so a browser with a dead cookie can still clean up.",
        "operationId": "logout",
        "responses": {
          "204": {
            "description": "No Content",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "apiKey": []
          },
          {
            "session": []
          }
        ],
        "summary": "Sign out",
        "tags": [
          "auth"
        ],
        "x-simhook-scope": "session"
      }
    },
...
    "/v1/auth/password-reset": {
      "post": {
        "description": "Completes a reset. Every existing session is signed out.",
        "operationId": "reset-password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Set a new password with a reset code",
        "tags": [
          "auth"
        ]
      }
    },
    "/v1/auth/password-reset/request": {
      "post": {
        "description": "Emails a code if the address has an account. The response is the same either way.",
        "operationId": "request-password-reset",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetRequestInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Request a password reset code",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/v1/auth/register": {
      "post": {
        "description": "Creates the account, signs it in with the session cookies, and emails a verification code. Sending is blocked until the email is verified.",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOutputBody"
                }
              }
            },
            "description": "Created",
            "headers": {
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Create an account",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/v1/auth/verify-email": {
      "post": {
        "operationId": "verify-email",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOutputBody"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "security": [
          {
            "apiKey": []
          },
          {
            "session": []
          }
        ],
        "summary": "Verify the email address",
        "tags": [
          "auth"
        ],
        "x-simhook-scope": "session"
      }
    },
...
    "/v1/device/pair": {
      "post": {
        "description": "Exchanges a pairing code for a device record and a device token. Called by the phone app.",
        "operationId": "pair-device",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PairInputBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PairOutputBody"
                }
              }
            },
            "description": "Created"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIError"
                }
              }
            },
            "description": "Error"
          }
        },
        "summary": "Pair a phone",
        "tags": [
          "device"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://api.simhook.dev"
    }
  ]
}